requirements.txt
openai-whisper
ollama
vosk
pyaudio
pyttsx3
sounddevice
numpy